Authenticate the api

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -13,6 +13,11 @@ class AttendeeController extends Controller
     /**
      * Display a listing of the resource.
      */
+
+     public function __construct(){
+        $this->middleware('auth:sanctum')->except(['index', 'show', 'update']);
+    }
+
     public function index(Event $event)
     {
         $attendees = $event->attendees()->latest();
@@ -54,8 +59,9 @@ class AttendeeController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $event, Attendee $attendee)
+    public function destroy(Event $event, Attendee $attendee)
     {
+        $this->authorize('delete-event', [$event, $attendee]);
         $attendee->delete();
         return response(status: 402);
     }
